mentora-backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.0
httpx==0.27.2
aiofiles==24.1.0

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Document Parsing ──────────────────────────────────────────────────────────
pymupdf==1.24.11
pytesseract==0.3.13
python-pptx==1.0.2
python-docx==1.1.2
Pillow==10.4.0
pandas>=2.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
striprtf>=0.0.26
ebooklib>=0.18

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.3

# ── Cache ─────────────────────────────────────────────────────────────────────
redis==5.1.1

# ── Audio TTS (Coqui) ─────────────────────────────────────────────────────────
# TTS requires Python <3.12. On Python 3.12+, install the maintained fork manually:
#   pip install coqui-tts
# The backend works without TTS — audio features will return empty placeholder files.
# Uncomment the line below if you're on Python 3.9-3.11:
# TTS==0.22.0

# ── Audio Transcription (faster-whisper — Python 3.12 compatible) ─────────────
faster-whisper>=1.0.0

# ── NLP for Planner ───────────────────────────────────────────────────────────
spacy>=3.7.0
# spaCy English model — installed automatically via pip
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
dateparser>=1.1.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Utilities ─────────────────────────────────────────────────────────────────
pydantic==2.9.2
email-validator>=2.0.0
pydantic-settings==2.5.2
python-slugify==8.0.4
